design in progress

+++ src/main/java/com/pg/edu/api/data/TrainingData.java
@@ -2,11 +2,11 @@ package com.pg.edu.api.data;
 
 import java.util.List;
 
-public interface TrainingData extends Data{
+public interface TrainingData extends Data {
 
     //TODO should be configurable
-    List<Integer> getInputs();
+    List<Double> getInputs();
 
     //TODO change to appropriate type
-    List<Integer> getOutputs();
+    List<Double> getOutputs();
 }
